refactor: move command
